Revert "ignore "div" tag if there is no attributes attached to it"
This commit breaks rendering of <div> when inside the tag other
formatting tags are present. A newline is expected but nothing is
rendered instead. Not sure how to fix this properly, and do not
understand how browsers are doing it correctly, as I can't see why one
case gets a <br> and one not. Have to dig into HTML specifications,
I think...
This reverts commit 5672c9a2fb598440a36bc33045a011e8a7fb2385.